requirements.txt
django
djangorestframework
django-cors-headers

opencv-python-headless

numpy
Pillow

torch
torchvision
ultralytics

django-redis
redis

psutil
pandas
requests
seaborn
tqdm
matplotlib
openpyxl

psycopg2-binary
python-dotenv
dj-database-url

# Production server
gunicorn
whitenoise
